Re #15947 Alterations made to fix GNU Make + C compiler warnings

+++ Framework/Algorithms/test/SofQWTest.h
@@ -115,8 +115,8 @@ public:
         SofQWTest::runSQW<Mantid::Algorithms::SofQW>("NormalisedPolygon");
     bool nanFound = false;
 
-    for (int i = 0; i < result->getNumberHistograms(); i++) {
-      if (nanFound = isNanInHistogram(*result, i)) {
+    for (size_t i = 0; i < result->getNumberHistograms(); i++) {
+      if ((nanFound = isNanInHistogram(*result, i))) {
         break; // NaN found in workspace, no need to keep searching
       }
     }
@@ -137,7 +137,7 @@ private:
   bool isNanInHistogram(const Mantid::API::MatrixWorkspace &result,
                         const size_t index) {
     bool nanFound = false;
-    for (int i = 0; i < result.blocksize(); i++) {
+    for (size_t i = 0; i < result.blocksize(); i++) {
       if (isnan(result.y(index)[i]) || isnan(result.e(index)[i])) {
         nanFound = true;
         break; // NaN found in histogram, no need to keep searching
